feat(util): add util crate  (#16)
This branch adds the beginnings of a `mycelium-util` crate, which is
sort of like a mini `no-std` standard library. In particular, this came
out of wanting an `Error` trait to use in the hal experiments I'm
working on, and not being able to use the `std::error::Error` trait. I
thought putting our error trait in a separate crate from `hal-core` was
a good idea, so that we could use it elsewhere in the kernel.

I also got bored and wrote a quick spinlock. The spinlock is tested
using `loom`, and there's a `loom` feature flag to turn on so that the
spinlock can also participate in other crates' loom tests.

Signed-off-by: Eliza Weisman <eliza@elizas.website>