feat(x86_64): a kernel that boots (#27)
closes #26 

this branch adds everything we need to boot the kernel on x86_64, plus
an attempt at organizing the main kernel crate and modules. it adds
stubs of a kernel main function, which currently just prints the version
and architecture. additionally, i added a quick VGA driver
implementation for x86, so we can print stuff.

the x86_64 kernel uses the [`bootloader`] crate by @phil-opp, which
implements a simple bootloader in rust for x86. i thought about using
GRUB, but this is simpler and lets us implement the whole build as 
`cargo` commands, which is nice — it should be quite simple to set
up a dev environment. i'm not sure if this bootloader works for 32-bit
x86 targets or if it's x64-specific, so we may need to figure something
else out to target 32-bit platforms.

sorry if any of the code or docs added in this branch doesn't make sense,
i have the flu and am kinda delirious.

[`bootloader`]: https://github.com/rust-osdev/bootloader

<img width="832" alt="Screen Shot 2019-12-12 at 3 27 42 PM" src="https://user-images.githubusercontent.com/2796466/70757407-5280eb80-1cf4-11ea-8f23-1e44701510b3.png">