feat(clippy-ci): fix clippy running on ci (#66)
The `actions-rs/toolchain` action runs `rustup show` before installing the toolchain requested. Unfortunately, when `rustup show` is run, `rustup` will read the `rust-toolchain` file from the root of the project, and immediately download and install the toolchain mentioned there.

This means that the `nightly` toolchain is installed with no components specified, and can select a version which does not currently have the `clippy` component. As `rustup` will refuse to downgrade your rust nightly version, this leads to an incompatible version being installed, and tests failing.

This branch moves the toolchain installing step before the `actions/checkout` step, meaning that `rustup` can't see the `rust-toolchain` file, and doesn't install the incorrect toolchain version.