fix(x86_64): forcibly unlock COM1 when oopsing
Currently, the kernel oops handler will forcibly unlock the VGA buffer
before writing the oops message. Ths prevents deadlocks if an oops
occurs while it was locked. However, we also write traces to COM1, but
we don't unlock it when oopsing.

This commit fixes this, by ensuring we forcibly unlock both VGA and
COM1. I added a new function to forcibly unlock both VGA and COM1, and
made both the oops handler and fault ISRs call it. It was necessary to
make the fault ISRs unlock as well, since they also currently emit
traces _before_ calling into oops. This means that if an oops occurs due
to a fault, we will unlock VGA and COM1 twice. But...who gives a shit.

Signed-off-by: Eliza Weisman <eliza@buoyant.io>