fix(x86_64): forcibly unlock COM1 when oopsing (#72)
Currently, the kernel oops handler will forcibly unlock the VGA buffer
before writing the oops message. Ths prevents deadlocks if an oops
occurs while it was locked. However, we also write traces to COM1, but
we don't unlock it when oopsing.

This commit fixes this, by ensuring we forcibly unlock both VGA and
COM1 when oopsing. I also refactored the x64 ISRs a bit so that the
interrupt context for faults is printed by the oops function, rather than
by the ISR. This means we don't have to worry about force-unlocking
in multiple places.

Signed-off-by: Eliza Weisman <eliza@buoyant.io>