feat(x86_64): initial paging implementation (#71)
This PR adds a very basic initial implementation of a paging subsystem
for x86_64. We can now pass some very simple tests for translating an
identity mapped address to itself, and remapping a page. There is a
bunch of stuff that will need to be implemented in follow-up branches.

The `hal-x86_64` paging module uses recursive mapping, and currently
relies on the 511th PML4 entry being the recursive entry. This is 
*not* what the `bootloader` crate's "recursive_page_table" feature 
does...so we don't use that feature. Instead, we just instruct the 
bootloader to map the physical memory, and we set up our own recursive
mapping in entry 511. 

The advantages of hard-coding the recursive entry are:

* The implementation is much simpler
* The recursive address doesn't need to be stored and passed around, 
  which probably has at least a small perf benefit.
* 511 is all 1s in binary, and is octal 0o777, so it is _really_ 
  obvious when printing ddresses in octal (and pretty obvious when 
  printing them in hex, too!)

However, this might not be great in the future if we want to work with
other bootloaders, etc., so we can revisit it in the future. It works
for now.

There is a *lot* of work remaining, but I would really prefer to merge
this PR and implement missing functionality in follow-ups. This is a
pretty huge chunk of code already and I'd like to get it merged back.

Signed-off-by: Eliza Weisman <eliza@elizas.website>
Co-authored-by: Nika Layzell <nika@thelayzells.com>