feat(hal): allow dynamically sized pages
Currently, the `Size` trait only allows representing statically known
sizes. This interface makes sense when _mapping_ pages, since the users
of that API will likely be saying things like "I want to map this 4kb
virtual page to a 4kb physical page". However, it makes less sense when
_translating_ pages/addresses. If I ask the page table "what page is
this address on?", I currently have to know in advance the size of the
page, which doesn't make sense — that's part of the information I would
be trying to _query_ by asking for the page containing an address. The
current trait does not model this nicely.

This commit updates the `Size` trait in `hal_core::mem::Page` to allow
representing dynamically sized pages, with a `Size::size()` fn rather
than an associated constant. The previous constant-only `Size` trait is
now a new trait, `StaticSize`, which _implements_ `Size`; this allows
code which requires a statically known size to bound type parameters
with `StaticSize` rather than `Size`.

Unfortunately, `size` fields on structs can no longer be `PhantomData`,
as they may in some cases be...real data; however, the current set of
types implementing `StaticSize` are still all zero-sized so this
shouldn't matter _that_ much.

Signed-off-by: Eliza Weisman <eliza@elizas.website>