feat(util): add `InitOnce` and `Lazy` cells in `sync` (#87)
This branch adds new `InitOnce<T>` and `Lazy<T>` cells in
`mycelium_util::sync`, to model statics that must be initialized at
runtime. `InitOnce` is manually initialized to a value, while `Lazy` is
constructed with an initializer. For particularly performance-sensitive
use cases, `InitOnce` also offers an unsafe `get_unchecked` method,
which assumes that the cell has already been uninitialized, and elides
the checks. This is, of course, quite unsafe.

Signed-off-by: Eliza Weisman <eliza@elizas.website>