feat(alloc): buddy block page frame allocator that kinda sorta works (#86)
this branch adds a working page frame allocator based on the buddy-block
allocation algorithm. it seems to more or less work, and can allocate
both single pages and ranges of pages.

however, it doesn't currently ensure correct alignments for page sizes
larger than the architecture's minimum page size --- there's no
difference, currently, between allocating one 2MB page and allocating a
2MB range of 4KB pages. for the 2MB page to be usable, though, it would
have to be aligned on a 2MB boundary, while the 2MB range of 4KB pages
only needs to start on a 4K boundary.

i'd like to add that in a future PR. for now, it just returns an error
when allocating page sizes that would require larger alignments, which
is fine for now i guess.

there's probably additional cleanup we could do, but whatever, i'm tired
of working on this and i want to merge it. we can make it better later.

a huge thank you to @mystor for help catching a couple of really
silly mistakes i made and getting this over the finish line!

Signed-off-by: Eliza Weisman <eliza@elizas.website>
Co-authored-by: Nika Layzell <nika@thelayzells.com>