feat(util): permit linked lists to be on the stack
Adding an additional associated type to the `Linked` trait type for the
node itself permits constructing short-lived linked lists on the stack,
using `Pin`  as the handle type. This may be useful in the future, if we
start using async/await for kernel tasks, and is also helpful for
writing tests for the linked list.

Signed-off-by: Eliza Weisman <eliza@buoyant.io>