chore: fix the build on the latest nightly, CI improvement (#102)
the latest Rust nightly breaks mycelium's build in a few different ways.
the `const_fn` feature gate was split up into several more granular
feature flags, so we need to enable the specific one for the const fn
features we're using. also, clippy added an enormous pile of horrible
new lints, so we need to fix that as well.

to prevent build breakage due to new nightlies in the future, i also
changed our build setup to pin to a specific nightly in
`rust-toolchain`, instead of just saying "nightly". CI now builds
against *that* nightly. i've added a new CI job to check the latest
nightly as an advisory; this will be allowed to fail, but is used as a
warning that updating will require fixing nightly breakage. it would be
nice to automate this completely with a scheduled action that checks if
we can build on the latest rust nightly, and automatically bumps the
toolchain if it works, but i didn't want to figure that out right now.
maybe a fun follow-up project though.

i also split out the warnings CI jobs that are allowed to fail into a
separate file with the name "CI (warnings)" rather than "CI". this
should make it clearer which jobs are required and which are allowed to
fail.

finally, i streamlined the CI flow a bit by removing the separate 
`cargo check` cross-compiling task. may as well just actually try the
build so everything can run in parallel.

Signed-off-by: Eliza Weisman <eliza@buoyant.io>