chore: add basic nix-shell stuff (#103)
This basically just makes sure you have QEMU and rustup.

It would be nice to also get it to install all the required
cargo install tools automagically, but I'll mess with that later.

Just installing rustup should ensure that the required rustup
components are present, since they're specified in the toolchain
file. Maybe there is also a more nix-y way to do that part, but
I dgaf.