feat(util): add packing pairs, other bit packing improvements
This commit adds new "packing pair" APIs to `util::bits` to allow
packing from a bit range in one value to a bit range in a different
value (at a different offset), without shifting to the least significant
bit and back.

This does still perform two shifts (a shift right and a shift left).
However, one of those shifts will always be by 0 bits. When the packing
pair is a const, we expect that the release mode compiler will hopefully
eliminate the zero-bit shift. When the pair is _not_ a const, two shifts
is probably more efficient than storing a bool for whether we should
perform a shift left or a shift right and branching on it.

I also cleaned up some stuff in the bit packing code.