chore: switch to Rust 2018 style modules (#116)
currently, we use Rust 2015 style modules --- if a module has submodules
in separate files, the main module is defined in a `mod.rs` file in a
directory with that module's name. this causes unnecessary git diff
churn if submodule files are added to a module that currently has no
submodules in separate files, since the file has to be renamed to
`mod.rs` and moved into a directory.

switching to Rust 2018 modules, where the main module file is named
`<MODNAME>.rs` and lives in the containing module's directory, and
submodule files live in the `<MODNAME>` directory, makes it easier to
add new submodule files without creating a big git diff where the whole
module was "added". however, doing that only for new code is kind of
weird, since all existing code uses Rust 2015 modules.

this PR just changes all existing modules to use the Rust 2018 style. no
actual code changes.