feat(inoculate): add subcommand for launching gdb (#119)
This branch adds a separate subcommand to `inoculate` for launching
`gdb` and connecting to a running QEMU on the default port. This is
intended for when the QEMU has already been launched, and you want to
connect to it a second terminal. I've found this useful for running a
debug session while looking at serial output in the QEMU terminal.

Signed-off-by: Eliza Weisman <eliza@buoyant.io>