feat(kernel): use buddy block allocator for kernel malloc (#130)
## Motivation

Currently, mycelium contains two separate memory allocators: a bump
pointer allocator which is used as the kernel's heap allocator, and a
buddy-block allocator which is used to allocate physical pages for the
paging subsystem.

This is not ideal. The bump pointer allocator was intended primarily as
an interim solution until a better allocation scheme can be implemented
(and, secondarily, for potential use in early kernel initialization
before we can cut over to a proper allocator). It has a few significant
limitations: currently, it is limited in size to 640 KB, and (more
importantly) it cannot reclaim freed allocations. So, once those 640 KB
are used up, we are --- to use a technical term here --- "totally
fucked". If we want to make a number of heap allocations over the
kernel's lifetime, such as for an (eventual) async task executor, we
need a kernel heap that supports reclaiming deallocated memory.

## Solution

This branch changes mycelium to use the buddy-block allocator both for
allocating page frames *and* for allocating kernel heap objects. The
bump pointer allocator is no longer used.

In the process of switching over to the buddy-block allocator, I
uncovered a number of bugs in the allocator implementation which didn't
occur when the allocator was only used to allocate 4KB pages. In
particular

* The buddy-block allocator will allow constructing a heap where the
  minimum region size is too low to fit all the free block metadata.
  This is uh, quite bad, since it means that some free block metadata
  may become corrupted if an adjacent region is allocated. So, I've
  fixed this by having the allocator round up the minimum allocated size
  to ensure the free block metadata fits in it.

* There were a couple assertions in the intrusive linked list and in the
  allocator that were just flat-out wrong. I've fixed and/or removed
  these.

* When freeing an allocation, the free-block metadata was incorrectly
  constructed with the size of the _requested allocation_, not the size
  of the block that was allocated for that allocation request. These may
  differ, since the buddy block allocator only allocates power-of-two
  sized blocks, and an allocation request might not be a power of two.
  Of course, this wasn't a problem when only allocating 4 KB page
  frames, but it quickly became a problem when trying to malloc random
  kernel objects.

* The use of `tracing` in the buddy-block allocator is problematic when
  using it as the kernel's heap. When a `tracing` span or event is hit
  for the first time, it will register the callsite in the global
  callsite registry, which is a `Vec`. If there's not enough space in
  the callsite registry for the new callsite, it may reallocate. When
  the event that caused the reallocation is inside of malloc...well, you
  can imagine what happens next. In order to fix this, I've switched
  mycelium to use a Git dependency on `tracing` 0.2.x. In 0.2,
  `tracing-core`'s global callsite registry was replaced with an atomic
  linked list implemented in the callsite statics, so the global
  callsite registry no longer allocates. In fact, mycelium is now able
  to use `tracing` without any allocations.

I also added some code for dumping the current heap state to the serial
port when the kernel oopses. This should help to debug heap-related
panics.

## Future Work

This is *also* kind of an interim solution. In the long term, we
probably *don't* actually want to use the same buddy block allocator to
allocate both page frames and kernel heap objects. Allocating small
kernel heap objects at the same level as 4 KB (and 2 MB or 1 GB) page
frames can cause memory fragmentation over long runtimes, which is not
ideal.

Instead, we probably want to use a tiered allocation scheme, where the
heap requests an entire physical page from the page allocator, and
performs allocations within that page. When all of the currently
allocated pages are full, the kernel heap would request more pages from
the page allocator. When all the allocations on a given page are
deallocated, that page might be released back to the page allocator, so
that it can then be allocated as a page frame _or_ as a new kernel heap
page.

However, this is more complex than just using one allocator for
everything, and I wanted to get to the point where we could reasonably
use malloc in the kernel. In follow-up changes, I intend to implement a
new slab allocator for heap objects that requests pages from the page
allocator.

Another advantage of a tiered approach is that pages given to the kernel
heap allocator can be mapped to arbitrary addresses in the virtual
address space. This would allow us to say things like, "the kernel heap
starts at `0xFFFF_FFFF_0000_0000`" or something, so that heap pointers can
be immediately identified by their addresses. But, the approach in this
branch at least lets us malloc things, so I'm calling that a significant
improvement.

Closes #19

Signed-off-by: Eliza Weisman <eliza@elizas.website>