chore(ci): move latest-nightly check to a cron job (#135)
This changes the GitHub Actions job that checks if mycelium builds on the
latest Rust nightly from a PR workflow to a cron job. This way, we don't
get the red X on PRs, because it's almost never the PR's fault that
nightly broke. Also, we won't miss potential nightly breakage if no
one's committing.

In the future, it could be nice to make this workflow automatically
update the pinned nightly if everything worked successfully. But, that
seems like a bit of work...

Signed-off-by: Eliza Weisman <eliza@buoyant.io>