change `Linked` trait to be generic over `Links`
This allows an individual type to be linked into multiple types of
intrusive data structure. For example, we might want to have async tasks
be part of both a run queue _and_ a global list of active tasks for
cancellation, so they would need separate sets of links for those two
structures.

A given instance of a type may still only be in a single intrusive
structure of each type at a time, though.