cut `intrusive::mpsc_queue::loom::mpmc` test in half
Turns out this test takes a _really_ long time to run: 78,602,709
iterations! This takes over an hour to complete on my machine, and
significantly longer on GitHub Actions (the last commit appears to be
still running :/).

This commit reduces the number of elements enqueued by each thread in
the test by half. We don't actually need to enqueue that many elements
to test the behavior we're trying to test here; loom will ensure that
all the possible paths through the memory model are simulated, so we
don't need to double the amount of times we hit any particular path of
code. This reduces the model's runtime _significantly_; it now only
takes 100,461 iterations and completes in 2 seconds (on my machine).