feat(cordyceps): add lock-free intrusive MPSC queue (#136)
This branch adds an intrusive lock-free MPSC queue implementation. The
queue is based on [Dmitry Vyukov's intrusive lock-free MPSC queue][1]
from 1024cores.net. This is primarily intended for use in an upcoming
async runtime for a queue of notified tasks for each runtime worker, but
it may be useful for other things, as well (possibly in the allocator,
though I'm not sure if it'll work there or not).

The new lock-free MPSC queue is added in a new crate called `cordyceps`,
which factors out the previous `mycelium_util::intrusive` module. This
is so that the intrusive collections library can be published to
crates.io, without making users depend on a bunch of mycelium-specific
code.

In order to implement this, I made some changes to the `list::Linked`
trait so that it could be used for multiple types of intrusive data
structure.

[1]: https://www.1024cores.net/home/lock-free-algorithms/queues/intrusive-mpsc-node-based-queue

Signed-off-by: Eliza Weisman <eliza@buoyant.io>