chore: bump nightly to 2022-02-15 (#138)
This branch updates Mycelium to build on Rust nightly 2022-02-15. This
appears to be the latest nightly that currently builds Mycelium; later
nightlies appear to introduce some weird issue where `rust-lld` attempts
to link with `libstdc++.so` and everything blows up. I don't super want
to figure that out right now.

In addition, this PR makes the following changes:

+ Update the use of the `asm!` macro to import it from `core::arch`,
  which is now required.
+ Update the `wasmi` dep's use of the `asm!` macro to `core::arch::asm!`
  (hawkw/wasmi@5caff845d5dbfdbdb82c624f470c55c71bfbe453)
+ Update the `bootloader` crate to v0.10.12, to ensure that it uses the
  `asm!` macro correctly

Closes #132

Signed-off-by: Eliza Weisman <eliza@elizas.website>