fix(cordyceps): add miri tests, fix stacked borrows (#139)
this branch adds a configuration for running Miri tests. currently, this
is primarily useful for testing `cordyceps` and friends, as the kernel
and HAL code contains a bunch of bare-metal stuff Miri definitely won't
understand.

after running Miri tests, i've fixed a number of stacked borrows
violations in the cordyceps linked list and MPSC queue modules. in
particular, i've changed the `Linked` trait a bit --- `Linked::as_ptr`
is now `Linked::into_ptr` and takes ownership of a `Handle`. this way,
we'll never produce an owning `Handle` (such as a `Box`) from a raw
pointer that originated from a shared reference (such as a
`&Pin<Box<...>>`). shared refs can still be used when the _handle type
itself_ is a shared ref. miri is now happy with us.