refac: move the test framework into `mycotest`
This PR moves the entire test framework implementation into the
`mycotest` crate, rather than having it be split across the
`mycelium-util` and `mycotest` crates. I also did some refactoring so
so that more of the test runner code can be defined in a cross-platform
way, rather than in the kernel's arch module.

Signed-off-by: Eliza Weisman <eliza@buoyant.io>