fix(x86_64): unfuck interrupt stack frames (#148)
It turns out that the reason we've been seeing extremely spooky behavior
in interrupt handlers, especially when trying to disassemble at `[rip]`,
is because LOL ELIZA FUCKED UP THE REGISTERS LOL LMAO.

When an interrupt occurs, a number of registers are pushed to the stack
frame for the interrupt handler. This includes the instruction pointer,
stack pointer, and the code and stack segment selectors. We defined a
`Registers` struct that consists of named fields for these registers, in
the correct order. However, because I'm a **loser fucking idiot**, the
current x86_64 ISRs were accidentally taking an `&mut Registers` rather
than a `Registers`. This meant we were interpreting the first 64 bits of
the registers (which I believe was the zero-padded 16-bit `ss` stack
segment selector) as a *pointer*, and reading the "register" values from
`[ss]`. This results in garbage.

And, when we tried to use `yaxpeax` to disassemble `[rip]`...`[rip +10]`
in the oops handler, we followed the "`rip`" from the location
pointed to by the stack segment selector, which was usually not pointing
at anything good. Thus, disassembly would always fault.

This branch fixes the ISR prototypes to take the registers by value, and
puts back `yaxpeax` disassembly.

Signed-off-by: Eliza Weisman <eliza@elizs.website>