fix(hal_core): never use multiline formatting for addrs (#150)
Since addrs are a tuple struct with a single small field, multiline
formatting never, ever looks good. Furthermore, since the trace
formatter switches to multiline mode when encountering a newline,
including a newline in the `Debug` impl for `PAddr`/`VAddr` means that
including one in an event will *always* put the logger in multiline
mode, even when every other field is a number or bool or something. That
results in a lot of unnecessary multiline logs, particularly in alloc.

This changes the `Debug` impls for addr types from using
`fmt::DebugTuple` to just do it manually, so it won't switch to
multiline mode when using the alt mode formatter.

Signed-off-by: Eliza Weisman <eliza@buoyant.io>