feat(mycotest): add assertion macros (#156)
Using `assert!` from `core` in `mycotest` tests is not great. The stdlib
assertions fail a test by panicking, which means that other tests will
not be permitted to run. This is kind of a shame as it means that fewer
test results will be reported --- when using `libtest` tests in code
that uses the stdlib, panics in tests are captured, so that multiple
tests can fail, but mycelium cannot do this. Instead, mycotests can
return a `Result` on failure, to allow other tests to run.

This branch adds new `assert!` and `assert_eq!` macros that return `Err`
when the assertion fails. These can be used in mycotests to not crash
the entire kernel when one test fails, allowing the outcome of other
tests to be reported even if the first one fails.

Signed-off-by: Eliza Weisman <eliza@buoyant.io>