feat(cordyceps): unsafe `MpscQueue` const constructor (#163)
In order to permit constructing a `cordyceps::MpscQueue` in a `static`,
this branch adds a new `const fn new_with_static_stub` to `MpscQueue`.
This takes an `&'static T` as the stub node. Because this would
potentially allow using the same stub node in multiple queues, this
function is unfortunately unsafe.

This branch also bumps the nightly version for `mycelium`, because
`cordyceps` and `mycelium-util` now require Rust 1.61.0, for trait
bounds in `const fn`s. The nightly bump required some changes to the
Miri tests for cordyceps, as Miri now seems to consume more memory than
it did previously.

Co-authored-by: Eliza Weisman <eliza@buoyant.io>