feat(async): implement task state transitions (#177)
this branch adds a proper implementation of task state transitions to
the async crate's task system. right now, the set of state transitions
is fairly small, but this will become more complex if we add
`JoinHandle`s or a similar abstraction for awaiting a task's completion.
also, i added tests exercising `wake_by_val` wakeups, and, it turns out,
we were leaking tasks in this case. so that's fixed now.

Signed-off-by: Eliza Weisman <eliza@buoyant.io>