feat(maitake): add `WaitQueue` (#191)
this branch adds a `WaitQueue` type in `maitake::wait`. this allows
multiple tasks to wait for an event, and be woken either one at a time
(in FIFO fashion), or all at once.

there are additional APIs we could add here, but this is a good starting
place. there probably should be docs examples but i'm lazy.

Signed-off-by: Eliza Weisman <eliza@buoyant.io>