fix(cordyceps): add `Drop` impl for `List` (#203)
When a linked list's `Handle` type owns the allocation for the item in
the list, it is necessary to drain all entries from the list and drop
them when the `List` instance is dropped. Otherwise, the entries may be
leaked.

This fixes a leak reported by Miri.

Fixes #165

BREAKING CHANGE:

The `List::new` constructor now requires a `T: Linked<list::Links<T>>`
bound.