feat(cordyceps): add `list::IterMut` (#208)
It doesn't really make sense to add a `DoubleEndedIterator` for the
`Cursor` type, because the cursor cannot iterate from the end of the
list --- it has a *single* position in the list. Similarly, it can't
actually be `ExactSizeIterator`, because it may remove entries while
iterating.

This commit adds a new `IterMut` for just iterating mutably.