fix(cordyceps): pin `list::IterMut` items (#209)
Currently, there is a potential safety issue when using
`cordyceps::list::IterMut`: the iterator's item type is `&mut T`.
`&mut T`s may be `mem::replace`d or `mem::take`en, moving them out of
the list to a different memory location. This would invalidate any
pointers in the list that point to that node.

Therefore, to ensure that the mutable iterator cannot corrupt the list
(potentially leading to dangling pointers or use-after-frees), this PR
changes `list::IterMut`'s `Item` type from `&'list mut T` to
`Pin<&'list mut T>`. This ensures that the iterator cannot be used to
corrupt the list.

Signed-off-by: Eliza Weisman <eliza@buoyant.io>