feat(maitake): add `wait::WaitMap`
This branch adds a `WaitMap` type, which allows for mailbox-like structure, usable for allowing waiters to receive data based on a specific key. This structure is somewhat similar to a `wait::WaitQueue`, however rather than being first-in first-out or wake-all based, it allows notifying specific wakers when data becomes available.