feat(maitake): add an async `Mutex` (#219)
This branch adds an async `Mutex` type to `maitake`, implemented using
`WaitQueue`. The implementation should be pretty straightforward,
although it required adding some additional methods to `WaitQueue`.
Currently, all the new `WaitQueue` functions are `pub(crate)`, but I'm
open to making them public APIs as well; I just wanted to save that for
a separate PR.

The `Mutex` type lives in a new `maitake::sync` module for
synchronization primitives --- I was on the fence about whether it still
makes sense to have a separate `maitake::wait` module if we are going to
have `maitake::sync`, but I didn't want to put the mutex in `wait`, as
it does more than *just* notify waiting tasks. Perhaps the `wait` types
should be moved into `sync`, but I'd prefer to do that in a separate PR.