test(maitake): fix tests with `default-features = false` (#222)
PR #219 added new tests for `maitake::sync::Mutex` that don't compile
when the "alloc" feature flag is not enabled. I forgot to make these
tests as `#[cfg(feature = "alloc")]`, so that PR broke the build with
`default-features = false`. Unfortunately, the CI job that tests with
`--no-default-features` was not marked as required, so although it
failed, PR #219 was still automerged and I didn't notice the build was
failing until it failed on `main`
(https://github.com/hawkw/mycelium/runs/6906013681?check_suite_focus=true)

This branch fixes the build failure by adding `cfg` attributes to the
tests that require the "alloc" feature. I also fixed some unused-imports
warnings in other tests that are only emitted with 
`default-features = false`.