feat(cordyceps): add `List::split_off`/`try_split_off` (#233)
This commit adds a `List::split_off` method to `cordyceps::List`,
similar to the one on `std::collections::LinkedList`. Additionally, it
adds a `try_split_off` method that doesn't panic if the split index is
greater than the list's length, and returns an `Option` instead ---
`std::collections::LinkedList` doesn't have this, but I thought it was
kinda nice.

Signed-off-by: Eliza Weisman <eliza@buoyant.io>