fix(cordyceps): pin `CursorMut` iterator `Item`s
Currently, the `Item` type of `list::CursorMut`'s `Iterator` impl is
`&mut T`. This is incorrect, as it allows the items to be moved out of
the list without unlinking them. Instead, the iterator's item type must
be `Pin`ned. PR #209 made this change for the `IterMut` type, but did
not make the same change for `CursorMut`, so this commit fixes
`CursorMut` similarly.

BREAKING CHANGE:

This changes the type signature of the `Iterator` impl for
`list::CursorMut`.