refac(maitake): store `WaitMap` data in `Wakeup`
This branch changes `WaitMap` waiters to store the received value in the
`Wakeup::DataReceived` variant. This removes the need for a separate
`UnsafeCell` in `WaitMap` waiters to store the data assigned to a
waiter, which lets us get rid of a little unsafe code, a `MaybeUninit`,
and also a `PinnedDrop` impl. This change makes a potential invalid
state where the waiter's state is advanced to `Wakeup::DataReceived`
without the `MaybeUninit` being initialized impossible to represent,
guarding against a possible read of uninitialized data.

Another side benefit is that this will slightly reduce the memory usage
of the `Waiter` node for `WaitMap`s. In the current code, the
`Wakeup` field uses`size_of::<Waker>() + size_of::<{enum descriminant}>()`
bytes, and the `val` field takes `size_of::<V>()` bytes. After this
change, we have a single field that's `max(size_of::<Waker>(),
size_of::<V>()) + size_of::<{enum descriminant}>()` bytes.