fix(alloc): fix wrong buddy block addr calculation (#247)
This fixes a bug in the buddy allocator where a block's buddy is
calculated wrongly. This resulted in an issue where we tried to remove a
block (which we *believed* was the current block's buddy) from the free
list for that order, but the block was actually on a different free
list.